Followup r100432
Disable challenge deletion


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@100436 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ specials/SpecialEditContest.php
@@ -312,7 +312,11 @@ class SpecialEditContest extends FormSpecialPage {
 		if ( $idString == '' ) {
 			return true;
 		}
-
+		global $wgContestDeletionEnabled;
+		if ( !$wgContestDeletionEnabled ) {
+			// Shouldn't get here (UI should prevent it)
+			throw new MWException( 'Contest deletion is disabled', 'contestdeletiondisabled' );
+		}
 		return ContestChallenge::s()->delete( array( 'id' => explode( '|', $idString ) ) );
 	}
